% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.score.R
\name{metric.score}
\alias{metric.score}
\title{Score metrics}
\usage{
metric.score(MetricName, MetricValue, IndexName, IndexRegion, Thresh)
}
\arguments{
\item{MetricName}{metric abbreviation}

\item{MetricValue}{calculated metric value}

\item{IndexName}{name of index (e.g., MBSS.2005.Bugs)}

\item{IndexRegion}{relevant bioregion or site class for given index}

\item{Thresh}{Data frame of Scoring Thresholds (Index.Name, Index.Region, Metric, Direction, Thresh.Lo, Thresh.Hi, ScoreRegime)}
}
\value{
vector of scores
}
\description{
This function calculates metric scores based on a Thresholds data frame.
Can generate scores for 1/3/5 (ScoreRegime="135") or 0-100 (ScoreRegime="100").
}
\examples{
## Load thresholds
thresh <- metrics_scoring
View(thresh)
## Define Index
IndexName <- "MBSS.2005.Bugs"

## Score Single Value
MetricName <- "nt_total"
MetricValue <- 18
IndexRegion <- "CP"
metric.score(MetricName,MetricValue,IndexName,IndexRegion,thresh)

## Score Data Frame of Values
# get metrics to score
myMetrics <- as.character(droplevels(unique(thresh[thresh[,"Index"]==IndexName,"Metric"])))
for (i in myMetrics){##FOR.i.START
  i.sc.nam <- paste0("Sc_",i)
  i.val <- Metrics.Bugs[,i]
  Metrics.Bugs[,i.sc.nam] <- metric.score(i,i.val,Metrics.Bugs$Index,Metrics.Bugs$Region,thresh)
  #
}##FOR.i.END
#
## View Results
View(Metrics.Bugs)
}
