% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.score.R
\name{metric.score}
\alias{metric.score}
\title{Score metrics}
\usage{
metric.score(MetricName, MetricValue, IndexName, IndexRegion, Thresh)
}
\arguments{
\item{MetricName}{metric abbreviation}

\item{MetricValue}{calculated metric value}

\item{IndexName}{name of index (e.g., MBSS.2005.Bugs)}

\item{IndexRegion}{relevant bioregion or site class for given index}

\item{Thresh}{Data frame of Scoring Thresholds (Index, Region, Metric, Direction, Thresh.Lo, Thresh.Hi, ScoreRegime)}
}
\value{
vector of scores
}
\description{
This function calculates metric scores based on a Thresholds data frame. f
Can generate scores for 1/3/5 (ScoreRegime="135") or 0-100 (ScoreRegime="100").
}
\examples{
thresh <- read.delim("metrics.scoring.tab")
MetricName <- "nt_total"
MetricValue <- 18
IndexName <- "MBSS.2005.Bugs"
IndexRegion <- "CP"

}

